requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch==2.2.0
torchvision==0.17.0
pillow==9.5.0
python-multipart
numpy==1.24.4